Add ledger and keystore tabs to unlock page
