+++ src/components/UnlockWithLedger.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function UnlockWithLedger({ onUnlockWallet }) {
+  return <div>Unlock with Ledger...</div>;
+}
+
+export default UnlockWithLedger;
